move resolveReference() to pointer.js
