+++ index.js
@@ -1,6 +1,6 @@
 const jaystring = require('jaystring')
 const genfun = require('./generate-function')
-const jsonpointer = require('./pointer')
+const { resolveReference } = require('./pointer')
 const formats = require('./formats')
 const KNOWN_KEYWORDS = require('./known-keywords')
 
@@ -8,47 +8,6 @@ const KNOWN_KEYWORDS = require('./known-keywords')
 const genobj = (name, property) => {
   if (!['string', 'number'].includes(typeof property)) throw new Error('Invalid property path')
   return `${name}[${JSON.stringify(property)}]`
-}
-
-const resolveReference = (root, additionalSchemas, ptr) => {
-  const visit = function(sub) {
-    if (sub && (sub.id || sub.$id) === ptr) return sub
-    if (typeof sub !== 'object' || !sub) return null
-    return Object.keys(sub).reduce(function(res, k) {
-      return res || visit(sub[k])
-    }, null)
-  }
-
-  const res = visit(root)
-  if (res) return res
-
-  ptr = ptr.replace(/^#/, '')
-  ptr = ptr.replace(/\/$/, '')
-
-  try {
-    return jsonpointer.get(root, decodeURI(ptr))
-  } catch (err) {
-    // do nothing
-  }
-
-  const end = ptr.indexOf('#')
-  // external reference
-  if (end === 0) {
-    return additionalSchemas[ptr]
-  } else if (end === -1) {
-    return additionalSchemas[ptr]
-  } else {
-    const ext = ptr.slice(0, end)
-    const fragment = ptr.slice(end).replace(/^#/, '')
-    try {
-      return jsonpointer.get(additionalSchemas[ext], fragment)
-    } catch (err) {
-      // do nothing
-    }
-  }
-
-  // falsy values will throw an error on usage
-  return null
 }
 
 const formatName = function(field) {
