+++ pointer.js
@@ -30,4 +30,45 @@ function get(obj, pointer) {
   return obj
 }
 
-module.exports = { get }
+function resolveReference(root, additionalSchemas, ptr) {
+  const visit = function(sub) {
+    if (sub && (sub.id || sub.$id) === ptr) return sub
+    if (typeof sub !== 'object' || !sub) return null
+    return Object.keys(sub).reduce(function(res, k) {
+      return res || visit(sub[k])
+    }, null)
+  }
+
+  const res = visit(root)
+  if (res) return res
+
+  ptr = ptr.replace(/^#/, '')
+  ptr = ptr.replace(/\/$/, '')
+
+  try {
+    return get(root, decodeURI(ptr))
+  } catch (err) {
+    // do nothing
+  }
+
+  const end = ptr.indexOf('#')
+  // external reference
+  if (end === 0) {
+    return additionalSchemas[ptr]
+  } else if (end === -1) {
+    return additionalSchemas[ptr]
+  } else {
+    const ext = ptr.slice(0, end)
+    const fragment = ptr.slice(end).replace(/^#/, '')
+    try {
+      return get(additionalSchemas[ext], fragment)
+    } catch (err) {
+      // do nothing
+    }
+  }
+
+  // falsy values will throw an error on usage
+  return null
+}
+
+module.exports = { get, resolveReference }
